lot of work

+++ SHREK/shrek.py
@@ -13,14 +13,18 @@ import os
 import sys
 from win32com.client import Dispatch
 import time
-import urllib
 from urllib.request import urlopen
+import pyautogui
+from pathlib import Path
 
 desktop = "%UserProfile%\\Desktop"
 
 path = "C:/Shrek_Files"
 url = "https://astrocore.net/gallery/meme.mp3"
 url2 = "https://astrocore.net/gallery/meme-ts1640990550.mp4"
+
+file1 = Path('C:/Shrek_Files/meme.mp4')
+file2 = Path('C:/Shrek_Files/meme.mp3')
 
 try:
     os.mkdir(path)
@@ -36,18 +40,22 @@ print("Preparing for launch.")
 video = "C:/Shrek_Files/meme.mp4"
 sound = "C:/Shrek_Files/meme.mp3"
 
-with open("C:/Shrek_Files/meme.mp3", 'wb') as local:
-  with urlopen(url) as remote:
-    local.write(remote.read())
+if file2.exists():
+    print("File exists. Skipping Download")
+else:
+    with open("C:/Shrek_Files/meme.mp3", 'wb') as local:
+        with urlopen(url) as remote:
+            local.write(remote.read())
 
-with open("C:/Shrek_Files/meme.mp4", 'wb') as local:
-  with urlopen(url2) as remote:
-    local.write(remote.read())
+if file1.exists():
+    print("File exists. Skipping Download")
+else:
+    with open("C:/Shrek_Files/meme.mp4", 'wb') as local:
+        with urlopen(url2) as remote:
+            local.write(remote.read())
 
-time.sleep(1)
-print('Starting in 5')
-time.sleep(1)
-print('Starting in 4')
+width, height= pyautogui.size()
+
 time.sleep(1)
 print('Starting in 3')
 time.sleep(1)
@@ -56,7 +64,7 @@ time.sleep(1)
 print('Starting in 1')
 
 def play():
-  while True:
+ while True:
       playsound(sound)
 
 def volume():
@@ -64,9 +72,9 @@ def volume():
 
     while True:
      for i in range(1000):
-         keyboard.press(Key.media_volume_up)
-         keyboard.release(Key.media_volume_up)
-         time.sleep(0.05)
+            keyboard.press(Key.media_volume_up)
+            keyboard.release(Key.media_volume_up)
+            time.sleep(0.05)
      time.sleep(2)
 
 def shutdown():
@@ -88,13 +96,15 @@ threading.Thread(target=play).start()
 threading.Thread(target=volume).start()
 threading.Thread(target=emergency).start()
 
-player = tkvideo(video, my_label, loop = 1, size = (1280,720))
+player = tkvideo(video, my_label, loop = 1, size = (width, height))
 player.play()
 
 def on_closing():
-   player = tkvideo(video, my_label, loop = 1, size = (1280,720))
+   player = tkvideo(video, my_label, loop = 1, size = (width, height))
    player.play()
 
 root.protocol("WM_DELETE_WINDOW", on_closing)
+
+print("Close this console to shutdown Virus!")
 
 root.mainloop()
